+++ app/providers.tsx
@@ -1,5 +1,5 @@
 'use client'
-import { useContext, useRef, ReactNode } from 'react'
+import { useContext, useRef, ReactNode, createContext, useState, useCallback } from 'react'
 import { LayoutRouterContext } from 'next/dist/shared/lib/app-router-context'
 import { ThemeProvider } from "next-themes"
 import { useEffect } from "react";
@@ -40,6 +40,78 @@ function FrozenRouter({ children }: { children: ReactNode }) {
   )
 }
 
+interface AudioContextType {
+  playing: boolean;
+  setPlaying: React.Dispatch<React.SetStateAction<boolean>>;
+  audioData: React.MutableRefObject<AnalyserNode | null>;
+  startPlaying: () => void;
+  stopPlaying: () => void;
+}
+
+const AudioContext = createContext<AudioContextType | null>(null);
+
+export function useAudioContext() {
+  const context = useContext(AudioContext);
+  if (!context) {
+    // This error will be thrown if the component is not a child of AudioProvider.
+    throw new Error('useAudio must be used within an AudioProvider');
+  }
+  return context;
+}
+
+function AudioProvider({ children }: { children: ReactNode }) {
+  const audioElement = useRef<HTMLAudioElement>(null);
+  const audioData = useRef<AnalyserNode | null>(null);
+  const [playing, setPlaying] = useState(false)
+  const audioContext = useRef<AudioContext | null>(null);
+  const musicType = '/audio/raingurl.mp3'
+
+  const startPlaying = () => {
+    if (playing || !musicType) return
+    // if there is no audio element or audio context, create them
+    if (!audioContext.current) {
+      createAudioContext()
+    }
+    // audioData is created in createAudioContext
+    if (!audioElement.current || !audioData.current) return
+    // set the source of the audio element to the music type
+    audioElement.current.src = musicType
+    // 
+    audioElement.current.play()
+    setPlaying(true)
+  }
+
+  const stopPlaying = useCallback(() => {
+    if (!audioElement.current) return
+    audioElement.current.pause()
+    setPlaying(false)
+  }, [setPlaying])
+
+  const createAudioContext = () => {
+    audioContext.current = new window.AudioContext();
+    if (!audioElement.current) return
+    const source = audioContext.current.createMediaElementSource(audioElement.current)
+    const analyzer = audioContext.current.createAnalyser()
+    analyzer.fftSize = 2048
+    source.connect(audioContext.current.destination)
+    source.connect(analyzer)
+    audioData.current = analyzer
+  }
+  const value = {
+    playing,
+    setPlaying,
+    audioData,
+    startPlaying,
+    stopPlaying,
+  }
+  return (
+    <AudioContext.Provider value={value}>
+      {children}
+    <audio id="audio" ref={audioElement}></audio>
+    </AudioContext.Provider>
+  )
+}
+
 
 export function Providers({ children }) {
   const pathname = usePathname();
@@ -67,7 +139,9 @@ export function Providers({ children }) {
       >
         <FrozenRouter>
           <ThemeProvider>
+            <AudioProvider>
             {children}
+            </AudioProvider>
           </ThemeProvider>
         </FrozenRouter>
       </motion.div>
