+++ components/default-layout.tsx
@@ -3,6 +3,7 @@ import { useSelectedLayoutSegment } from 'next/navigation'
 import styles from '../styles/layout.module.scss'
 import DefaultHeader from './default-header'
 import HeaderControls from './header-controls';
+import BackgroundAnimation from './background-animation';
 
 function DefaultLayout({ children }) {
   const segment = useSelectedLayoutSegment();
@@ -17,12 +18,12 @@ function DefaultLayout({ children }) {
 
   }
   return (
-    <div className={styles.container}>
-      <Header />
-      <main className={styles.content} id="main-content">
-        {children}
-      </main>
-    </div>
+    <BackgroundAnimation className={styles.container}>
+        <Header />
+        <main className={styles.content} id="main-content">
+          {children}
+        </main>
+    </BackgroundAnimation>
   )
 }
