+++ styles/utils.module.scss
@@ -148,7 +148,7 @@
     left: 0;
     height: 100%;
     width: 100%;
-    background-image: var(--noise-1), var(--bg-gradient);
+    // background-image: var(--noise-1), var(--bg-gradient);
     background-blend-mode: darken;
     // filter: blur(32px);
     z-index: -1;
